requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# requirements.txt
# These are all the Python libraries your backend needs.
# Render reads this file automatically and installs everything for you.
# On your laptop, install them with: pip install -r requirements.txt
# ─────────────────────────────────────────────────────────────────────────────

# Web server & API framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0       # ASGI server — runs your FastAPI app
python-multipart>=0.0.9         # Required for file uploads in FastAPI

# Image processing
Pillow>=10.0.0                  # Open, read, and manipulate images
piexif>=1.1.3                   # Read and write EXIF metadata
numpy>=1.26.0                   # Math/array operations for forensics

# HTTP requests (for calling Hugging Face API)
requests>=2.31.0

